Added animations and parallax scrolling

+++ source/repos/Sproutfall/Sproutfall/Reticle.cpp
@@ -24,23 +24,20 @@ Reticle::~Reticle()
 }
 void Reticle::Update(float tf)
 {
-	double xpos = m_window->mapPixelToCoords(sf::Mouse::getPosition(*m_window)).x - m_playerSprite->getPosition().x;
-	double ypos = m_window->mapPixelToCoords(sf::Mouse::getPosition(*m_window)).y - m_playerSprite->getPosition().y;
+	float xpos = m_window->mapPixelToCoords(sf::Mouse::getPosition(*m_window)).x - m_playerSprite->getPosition().x;
+	float ypos = m_window->mapPixelToCoords(sf::Mouse::getPosition(*m_window)).y - m_playerSprite->getPosition().y;
 	float scalingFactor = m_distanceFromPlayer;
 	if (xpos != 0)
 	{
-		double rads = atan(ypos / xpos);
-		double rotationMod = 0;
+		float rads = atan(ypos / xpos);
+		float rotationMod = 0;
 		if (xpos < 0)
 		{
 			scalingFactor = scalingFactor * -1;
 			rotationMod = 180;
 		}
-		double tmpx = m_playerSprite->getPosition().x + (scalingFactor * cos(rads));
-		double tmpy = m_playerSprite->getPosition().y + (scalingFactor * sin(rads));
-		//cout << tmpx << " " << tmpy << "\n";
-		//cout << cos(rads) << " " << sin(rads) << "\n";
-		//cout << rads << "\n";
+		float tmpx = m_playerSprite->getPosition().x + (scalingFactor * cos(rads));
+		float tmpy = m_playerSprite->getPosition().y + (scalingFactor * sin(rads));
 		m_Sprite->setPosition(tmpx, tmpy);
 		m_Sprite->setRotation(rotationMod + (rads * (180.0 / 3.141592653589793238463) + 90));
 	}
